+++ packages/basehub/src/react/code-block/client/index.ts
@@ -1,0 +1,2 @@
+export { useCodeBlockContext } from "./context";
+export { useCopyToClipboard } from "./use-copy-to-clipboard";