+++ packages/basehub/tsup-client.config.ts
@@ -8,6 +8,8 @@ export default defineConfig((_options: Options) => {
     dts: true,
     entry: {
       "react-rich-text": "./src/react/rich-text/index.ts",
+      "react-code-block/index": "./src/react/code-block/index.ts",
+      "react-code-block/client": "./src/react/code-block/client/index.ts",
       "react-search": "./src/react/search/index.ts",
       "api-transaction": "./src/api-transaction.ts",
       analytics: "./src/analytics/index.ts",
@@ -15,6 +17,8 @@ export default defineConfig((_options: Options) => {
       "next-image": "./src/next/image/index.ts",
     },
     format: ["esm"],
+    splitting: true,
+    clean: true,
     external: ["react", "react-dom", "next"],
     plugins: [
       {