+++ playground/CHANGELOG.md
@@ -1,5 +1,68 @@
 # playground
 
+## 0.0.128-code.8
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.2.4-code.8
+
+## 0.0.128-code.7
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.2.4-code.7
+
+## 0.0.128-code.6
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.2.4-code.6
+
+## 0.0.128-code.5
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.2.4-code.5
+
+## 0.0.128-code.4
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.2.4-code.4
+
+## 0.0.128-code.3
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.2.4-code.3
+
+## 0.0.128-code.2
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.2.4-code.2
+
+## 0.0.128-code.1
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.2.4-code.1
+
+## 0.0.128-code.0
+
+### Patch Changes
+
+- Updated dependencies
+  - basehub@7.2.4-code.0
+
 ## 0.0.127
 
 ### Patch Changes